Create 01-MathsQuiz.py
added a variable for background color and text color to make it easier when coding. defined RandomNumber as randint to generate random integers to be used in the maths quiz. defined Symbol to be randomised arithmetic operators. Created OOP and UI for the quiz, as well as a difficulty setting.