function for viewing all student information option
Defined a function called view_all_students. This is a button in the GUI interface. Its purpose is to provide a list of all the student informatio listed in the studentMarks.txt file and output it to Tkinter. IT also calculates the average percentage by using a formula with the total percentage divided by the length of the students variable